Clean up main file | Add port mapping flag and logic | port-map example
