+++ .gitignore
@@ -13,3 +13,5 @@
 
 # Build artifacts and other assets
 dist/
+
+awscurl
