+++ README.md
@@ -12,23 +12,6 @@ or [ECS Tasks](https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task-
 
 ## Installation
 
-### Cross-platform Binary (Recommended)
-
-Download appropriate version for your platform from [awscurl Releases](https://github.com/legal90/awscurl/releases).
-Once downloaded and unpacked, the binary can be run from anywhere. For convinience, you can place the binary to `/usr/local/bin/`
-
-### Docker
-awscurl container images are also released on DockerHub as [`legal90/awscurl`](https://hub.docker.com/repository/docker/legal90/awscurl)
-
-```
-$ docker run --rm \
-    --env AWS_ACCESS_KEY_ID \
-    --env AWS_SECRET_ACCESS_KEY \
-    legal90/awscurl --help   # see Usage section for other available arguments
-```
-_Note:_ This command assumes that you have `AWS_ACCESS_KEY_ID` and `AWS_SECRET_ACCESS_KEY` exported in your environment.
-You can find more details about AWS authentication in the _Usage_ section below.
-
 ### Build from Source
 
 #### Prerequisites
@@ -40,10 +23,11 @@ You can find more details about AWS authentication in the _Usage_ section below.
 
 `awscurl` uses the Go Modules support built into Go 1.11. The easiest way to get started is to clone awscurl in a directory
 outside of the $GOPATH, as in the following example:
+
 ```shell
 mkdir $HOME/src
 cd $HOME/src
-git clone https://github.com/legal90/awscurl.git
+git clone https://github.com/bkimmig/awscurl.git
 cd awscurl
 go build   # or `go install` to install it to your $GOPATH/bin
 ```
@@ -51,19 +35,21 @@ go build   # or `go install` to install it to your $GOPATH/bin
 ## Usage
 
 ```
-$ awscurl --help
+$ awscurl –help
 
 A simple CLI utility with cURL-like syntax allowing to send HTTP requests to AWS resources.
 It automatically adds Siganture Version 4 to the request. More details:
 https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html
 
 Usage:
-  awscurl [URL] [flags]
+  awscurl [flags]
 
 Flags:
       --access-key string      AWS Access Key ID to use for authentication
-  -d, --data string            Data payload to send within a POST request
-  -H, --header stringArray     Extra HTTP header to include in the request. Example: "Content-Type: application/json". Could be used multiple times
+  -d, --data string            Data payload to send within a request. Could be also read from a file if prefixed with @, example: -d "@/path/to/file.json"
+  -H, --header stringArray     Extra HTTP header to include in the request. Example: -h "Content-Type: application/json". Could be used multiple times
+  -k, --insecure               Allow insecure server connections when using SSL
+      --port-map string        map the local port to the remote port 'localport:remoteport'. Used in cases where you are locally port-forwarding the service
       --profile string         AWS awsProfile to use for authentication
       --region string          AWS region to use for the request
   -X, --request string         Custom request method to use (default "GET")
@@ -72,7 +58,8 @@ Flags:
       --session-token string   AWS Session Key to use for authentication
   -h, --help                   help for awscurl
   -v, --version                version for awscurl
-```
+
+
 
 ### AWS Authentication
 As you can see above, `awscurl` supports several argument, allowing you to choose the desired way to authenticate on AWS.
@@ -94,6 +81,31 @@ will follow "the default provider chain". It looks for credentials in this order
 3. IAM role for Amazon EC2 or Tasks (if you run `awscurl` on EC2 Instance or ECS task)
 
 ### Examples
+
+#### Requests to AWS Elasticsearch Service on a VPC
+
+This example signs requests from your local computer to a bastion server that is
+port-forwarding your Elasticsearch (ES) url.
+
+First you need to edit your `/etc/hosts` file to map `localhost` to your ES VPC.
+
+```
+localhost your-es-vpc-id-here
+```
+
+Then you can port-forward your service which will forward the port 443 to your
+localhost 9200
+
+`ssh -i ~/.ssh/your-key user@ip-address -L 9200:your-es-vpc-id-here.region.es.amazonaws.com:443`
+
+Given that and the edit to your hosts file you can now use the `awscurl` command
+to sign request through the bastion server - the port map says use the first
+port locally but sign the request with the second port.
+
+`awscurl --service es --region your-region --profile your-profile --port-map "9200:443" "https://your-vpc.region.es.amazonaws.com:9200/"`
+
+Which will sign the request for the port 443 but pass it to your 9200 port
+locally. This should return the basic info about your cluster.
 
 #### Call S3: List bucket content
