+++ go.mod
@@ -1,6 +1,6 @@
-module github.com/legal90/awscurl
+module github.com/bkimmig/awscurl
 
-go 1.16
+go 1.13
 
 require (
 	github.com/aws/aws-sdk-go-v2 v0.22.0
